W1252string in a separate .h file
