+++ modules/c++/coda-oss.vcxproj.filters
@@ -22,34 +22,34 @@
       <Filter>gsl</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\bit.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\CPlusPlus.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\cstddef.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\namespace_.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\optional.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\optional_.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\span.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\span_.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\string.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\type_traits.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="except\include\except\Backtrace.h">
       <Filter>except</Filter>
@@ -733,7 +733,7 @@
       <Filter>types</Filter>
     </ClInclude>
     <ClInclude Include="coda_oss\include\coda_oss\numbers.h">
-      <Filter>coda_oss</Filter>
+      <Filter>math\coda_oss</Filter>
     </ClInclude>
     <ClInclude Include="hdf5.lite\source\hdf5.lite.h">
       <Filter>hdf5.lite</Filter>
@@ -947,6 +947,9 @@
     </ClInclude>
     <ClInclude Include="mem\include\mem\AutoPtr.h">
       <Filter>mem</Filter>
+    </ClInclude>
+    <ClInclude Include="str\include\str\W1252string.h">
+      <Filter>str</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -1403,9 +1406,6 @@
     <Filter Include="gsl">
       <UniqueIdentifier>{619ad1da-f21c-4027-9b5d-23f08225b96f}</UniqueIdentifier>
     </Filter>
-    <Filter Include="coda_oss">
-      <UniqueIdentifier>{59f3d9a1-06d3-4779-aef2-cc55223c3017}</UniqueIdentifier>
-    </Filter>
     <Filter Include="std">
       <UniqueIdentifier>{3051f4b6-dad4-47ea-b4bd-d25d9e09f782}</UniqueIdentifier>
     </Filter>
@@ -1489,6 +1489,9 @@
     </Filter>
     <Filter Include="unique">
       <UniqueIdentifier>{79aa73a7-ed0d-4826-b7ed-c8d91d96998a}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="math\coda_oss">
+      <UniqueIdentifier>{59f3d9a1-06d3-4779-aef2-cc55223c3017}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
